Fix: silence sign-compare warnings in Clang build (Cluster H)
Seven sites across five files where the MIN() macro receives mixed
signed/unsigned integer arguments:

dkim-canon.c: MIN(size_t buflen, ssize_t canon_remain) — the
  surrounding guard already verifies canon_remain != -1, so it is
  non-negative; cast to size_t is safe.

opendkim-dns.c: MIN(size_t ubd_buflen, int answer_len) — answer_len
  is a packet length from libunbound and is always non-negative; cast
  to size_t is safe.

dkim.c (dkim_getsighdr_d): MIN(size_t margin-len, ptrdiff_t y-x) —
  the while-loop guard (x < y) guarantees y-x > 0; also change the
  accumulator n from int to size_t (it is only ever used as a byte
  count for dkim_dstring_catn and pointer arithmetic).

dkim.c (dkim_get_sigsubstring): MIN(size_t *buflen-1, int minsiglen)
  — minsiglen is set to x+1 where x is a non-negative loop index;
  add explicit (size_t) cast and (int) cast for the assignment to the
  existing int local minlen.

opendkim-genzone.c (×2):
  - MIN(off_t s.st_size, size_t *buflen): fstat succeeded, so
    st_size >= 0; cast to size_t is safe.
  - MIN(size_t sizeof(derdata), int outlen): outlen comes from
    BIO_get_mem_data after a successful BIO_flush; cast to size_t.

opendkim-testkey.c: same pattern as opendkim-genzone.c — fstat
  succeeded before the MIN call.

+++ libopendkim/dkim-canon.c
@@ -153,7 +153,7 @@ dkim_canon_write(DKIM_CANON *canon, const u_char *buf, size_t buflen)
 	assert(canon != NULL);
 
 	if (canon->canon_remain != (ssize_t) -1)
-		buflen = MIN(buflen, canon->canon_remain);
+		buflen = MIN(buflen, (size_t) canon->canon_remain);
 
 	canon->canon_wrote += buflen;
